refactored init as a plugin
