+++ plugins/init/package.json
@@ -1,0 +1,8 @@
+{
+    "name": "@reframe/init",
+    "version": "0.0.1-rc.12",
+    "publishConfig": {
+        "access": "public"
+    },
+    "repository": "github:reframejs/reframe"
+}
