+++ plugins/init/readme.md
