+++ plugins/init/templates/coreFilesTemplate.js
