+++ plugins/init/templates/homeTemplate.js
